Added "disabled" accessories to remove temporarily

+++ platform.js
@@ -69,7 +69,7 @@ const BroadlinkRMPlatform = class extends HomebridgePlatform {
     let tvs = [];
     config.accessories.forEach((accessory) => {
       if (!accessory.type) throw new Error(`Each accessory must be configured with a "type". e.g. "switch"`);
-
+      if (accessory.disabled) return;
       if (!classTypes[accessory.type]) throw new Error(`homebridge-broadlink-rm doesn't support accessories of type "${accessory.type}".`);
 
       const homeKitAccessory = new classTypes[accessory.type](log, accessory);
